requirements.txt
Django==3.1.6
psycopg2-binary==2.9.7
python-dotenv==0.15.0
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers==3.7.0
django-ses==1.0.3
requests==2.28.2
gunicorn==20.0.4
django-health-check==3.16.3
celery==5.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.21
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
pillow==9.0.0
python-dotenv==0.15.0
django-extensions==3.1.5
django-storages==1.12.3
django-s3-storage==0.13.4
sentry-sdk==1.17.0
redis==4.5.1